scoreboard fix & addtional config options

+++ src/main/java/me/ShermansWorld/raidsperregion/util/ScoreboardUtil.java
@@ -8,40 +8,38 @@ import fr.mrmicky.fastboard.FastBoard;
 import me.ShermansWorld.raidsperregion.raid.Raid;
 
 public class ScoreboardUtil {
-	
-	
+
 	public static FastBoard createNewScoreboard(Player player, Raid raid, String title) {
 		FastBoard scoreboard = new FastBoard(player);
 		scoreboard.updateTitle(title);
 		raid.getScoreboards().put(player.getUniqueId(), scoreboard);
 		return scoreboard;
 	}
-	
-	public static void updateBoard(FastBoard board, String ... lines) {
-		
+
+	public static void updateBoard(FastBoard board, String... lines) {
+
 		if (board == null || board.isDeleted()) {
 			return;
 		}
-		
+
 		// update colors
-		for (int i = 0; i < lines.length; ++i ) {
+		for (int i = 0; i < lines.length; ++i) {
 			lines[i] = Helper.color(lines[i]);
 		}
-		
+
 		board.updateLines(lines);
 	}
-	
+
 	public static FastBoard getScoreboard(Player player, Raid raid) {
 		return raid.getScoreboards().get(player.getUniqueId());
 	}
-	
+
 	public static void clearScoreboards(Raid raid) {
 		for (UUID uuid : raid.getScoreboards().keySet()) {
-			try {
+			if (raid.getScoreboards().get(uuid) != null) {
 				raid.getScoreboards().get(uuid).delete();
-			} catch(IllegalStateException e) {
 			}
 		}
 	}
-	
+
 }
